+++ system-test/src/main/java/io/github/galop_proxy/system_test/server/StatusCodeContextHandler.java
@@ -1,0 +1,12 @@
+package io.github.galop_proxy.system_test.server;
+
+import org.eclipse.jetty.server.handler.ContextHandler;
+
+final class StatusCodeContextHandler extends ContextHandler {
+
+    StatusCodeContextHandler() {
+        super("/status-code");
+        setHandler(new StatusCodeHandler());
+    }
+
+}